# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new transaction (supports expense, income, or transfer)",
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "EXPENSE",
                "INCOME",
                "TRANSFER"
              ],
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transactions with filtering",
        "tags": [
          "transactions"
        ]
      }
    },
    "/v1/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction details by ID",
        "tags": [
          "transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a transaction by ID",
        "tags": [
          "transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a transaction by ID",
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gustam Platform API",
    "description": "Central API for Gustam Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "EXPENSE",
            "description": "Transaction type",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "number",
            "example": 12.5,
            "description": "Amount of transaction"
          },
          "title": {
            "type": "string",
            "example": "Coffee",
            "description": "Title of transaction"
          },
          "description": {
            "type": "string",
            "example": "Venti Latte at Starbucks",
            "description": "Description of transaction"
          },
          "accountId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
            "description": "Account UUID"
          },
          "categoryId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a12",
            "description": "Category UUID (optional for income/transfers)"
          },
          "transferAccountId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a13",
            "description": "Transfer Destination Account UUID"
          },
          "category": {
            "type": "string",
            "example": "Food",
            "description": "Deprecated free-text category"
          },
          "date": {
            "type": "string",
            "example": "2026-07-15T00:00:00.000Z",
            "description": "Date"
          }
        },
        "required": [
          "type",
          "amount",
          "title",
          "accountId"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "EXPENSE",
            "description": "Transaction type",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "number",
            "example": 12.5,
            "description": "Amount of transaction"
          },
          "title": {
            "type": "string",
            "example": "Coffee",
            "description": "Title of transaction"
          },
          "description": {
            "type": "string",
            "example": "Venti Latte at Starbucks",
            "description": "Description of transaction"
          },
          "accountId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
            "description": "Account UUID"
          },
          "categoryId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a12",
            "description": "Category UUID"
          },
          "transferAccountId": {
            "type": "string",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a13",
            "description": "Transfer Destination Account UUID"
          },
          "category": {
            "type": "string",
            "example": "Food",
            "description": "Deprecated free-text category"
          },
          "date": {
            "type": "string",
            "example": "2026-07-15T00:00:00.000Z",
            "description": "Date"
          }
        }
      },
...
    }
  }
}